requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
certifi==2025.1.31
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
logging==0.4.9.6
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
redis==5.2.1
sniffio==1.3.1
typing-inspection==0.4.0
typing_extensions==4.13.2
websockets==15.0.1
